move ClientType from middleware to models
